Here it is! Project Finished!
I created an interface and implementations of it for entity framework or rawSQL depending on what the user chooses (weigths or cardio).

I did not create a method to create the database for rawSQL because i imagine that the migrations folder already takes care of that, right? So i though it was not needed.